src/ecommerce/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
boto3==1.34.0
gunicorn==21.2.0
